Cast idx to uint32 to fix compilation on i386 (#1118)
`env GOOS=linux GOARCH=386` go build no longer complains about integer overflow.

See https://github.com/dgraph-io/badger/issues/953#issuecomment-552241033

+++ table/table.go
@@ -432,7 +432,7 @@ func (t *Table) block(idx int) (*block, error) {
 
 func (t *Table) blockCacheKey(idx int) uint64 {
 	y.AssertTrue(t.ID() < math.MaxUint32)
-	y.AssertTrue(idx < math.MaxUint32)
+	y.AssertTrue(uint32(idx) < math.MaxUint32)
 	return (t.ID() << 32) | uint64(idx)
 }
